- Auto-generate all C/C++ translation code (issue #33). - Change index parameter types from int to size_t to make 0-based range implicit. - Make CefPrintOptions and CefMenuInfo proper wrapper classes. - Normalize the naming of menu-related types. - Remove unused command_line variable from test_suite.cc.
git-svn-id: https://chromiumembedded.googlecode.com/svn/trunk@408 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ libcef/v8_impl.h
@@ -127,14 +127,14 @@ public:
   virtual CefTime GetDateValue() OVERRIDE;
   virtual CefString GetStringValue() OVERRIDE;
   virtual bool HasValue(const CefString& key) OVERRIDE;
-  virtual bool HasValue(int index) OVERRIDE;
+  virtual bool HasValue(size_t index) OVERRIDE;
   virtual bool DeleteValue(const CefString& key) OVERRIDE;
-  virtual bool DeleteValue(int index) OVERRIDE;
+  virtual bool DeleteValue(size_t index) OVERRIDE;
   virtual CefRefPtr<CefV8Value> GetValue(const CefString& key) OVERRIDE;
-  virtual CefRefPtr<CefV8Value> GetValue(int index) OVERRIDE;
+  virtual CefRefPtr<CefV8Value> GetValue(size_t index) OVERRIDE;
   virtual bool SetValue(const CefString& key, CefRefPtr<CefV8Value> value,
                         PropertyAttribute attribute) OVERRIDE;
-  virtual bool SetValue(int index, CefRefPtr<CefV8Value> value) OVERRIDE;
+  virtual bool SetValue(size_t index, CefRefPtr<CefV8Value> value) OVERRIDE;
   virtual bool SetValue(const CefString& key, AccessControl settings, 
                         PropertyAttribute attribute) OVERRIDE;
   virtual bool GetKeys(std::vector<CefString>& keys) OVERRIDE;
